feat: add session loader component
