+++ src/components/providers/index.ts
@@ -1,8 +1,3 @@
 export { AuthProvider } from "./auth-provider";
 export { QueryProvider } from "./query-provider";
-export type {
-  SessionLoadingContextValue,
-  SessionLoadingProviderProps,
-  SessionLoadingState,
-} from "./session-loading-provider";
-export { SessionLoadingProvider, useSessionLoading } from "./session-loading-provider";
+export { SessionLoader } from "./session-loader";
